Move statistics into separate file
