+++ rustynum-rs/src/num_array/mod.rs
@@ -4,5 +4,6 @@ mod impl_clone_from;
 pub mod linalg;
 mod manipulation;
 pub mod operations;
+mod statistics;
 
 pub use array_struct::{NumArray, NumArrayF32, NumArrayF64, NumArrayI32, NumArrayI64, NumArrayU8};
